Usermode - Fix crash caused by invalid data size passed to ASM

+++ Usermode/libtifflin_syscalls/src/raw-amd64.rs
@@ -4,7 +4,7 @@
 			let rv;
 			asm!("syscall"
 				: "={rax}" (rv)
-				: "{rax}" ($id) $(, $reg ($val))*
+				: "{rax}" ($id as usize) $(, $reg ($val as usize))*
 				: "rcx", "r11"
 				: "volatile"
 				);
